change application root
