--- Dockerrun.aws.json
@@ -1,22 +1,0 @@
-{
-  "AWSEBDockerrunVersion": "1",
-  "Image": {
-    "Name": "970659526887.dkr.ecr.us-east-2.amazonaws.com/blacklist-service:latest",
-    "Update": "true"
-  },
-  "Ports": [
-    {
-      "ContainerPort": "5000"
-    }
-  ],
-  "Environment": [
-    {
-      "Name": "DATABASE_URL",
-      "Value": "postgresql://postgres:password@database-1.cr2u068mwzvv.us-east-2.rds.amazonaws.com:5432/blacklist_db"
-    },
-    {
-      "Name": "JWT_SECRET_KEY",
-      "Value": "supersecret"
-    }
-  ]
-}
