--- docker-compose.yml
@@ -1,29 +1,0 @@
-services:
-  blacklist_service:
-    build: .
-    ports:
-      - "80:5000"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - JWT_SECRET_KEY=${JWT_SECRET_KEY}
-    # environment:
-    #   - DATABASE_URL=postgresql://user:password@db:5432/blacklist_db
-    #   - JWT_SECRET_KEY=supersecret
-    # depends_on:
-    #   db:
-    #     condition: service_healthy
-
-    
-  # db:
-  #   image: postgres:13
-  #   environment:
-  #     POSTGRES_DB: blacklist_db
-  #     POSTGRES_USER: user
-  #     POSTGRES_PASSWORD: password
-  #   ports:
-  #     - "5433:5432"
-  #   healthcheck:
-  #     test: ["CMD", "pg_isready", "-U", "user"]
-  #     interval: 10s
-  #     timeout: 5s
-  #     retries: 5
